requirements.txt
streamlit==1.30.0
tensorflow==2.19.0
Pillow==10.0.0
numpy==1.24.3
